Домашнее задание 2. Пункт 2: добавлент Navigation Drawer.
