--- app/src/main/java/ru/lyubimov/weather/weatherapp/SingleFragmentActivity.java
@@ -1,34 +1,0 @@
-package ru.lyubimov.weather.weatherapp;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-
-/**
- * Created by Alex on 13.12.2017.
- */
-
-public abstract class SingleFragmentActivity extends AppCompatActivity {
-
-    protected abstract Fragment createFragment();
-
-    protected int getLayoutResId() {
-        return R.layout.activity_fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(getLayoutResId());
-        FragmentManager manager = getSupportFragmentManager();
-        Fragment fragment = manager.findFragmentById(R.id.fragment_container);
-
-        if (fragment == null) {
-            fragment = createFragment();
-            manager.beginTransaction()
-                    .add(R.id.fragment_container, fragment)
-                    .commit();
-        }
-    }
-}
