+++ app/src/main/res/drawable/side_nav_bar.xml
@@ -1,0 +1,9 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="135"
+        android:centerColor="#4c9baf"
+        android:endColor="#3777a4"
+        android:startColor="#7cceca"
+        android:type="linear" />
+</shape>
