+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <string name="app_name">WeatherApp</string>
+    <string name="web">dr_kuvento@mail.ru</string>
 
     <string name="no_location_detected">No location detected. Make sure location is enabled on the device.</string>
     <string name="permission_rationale">Location permission is needed for core functionality</string>
@@ -28,5 +29,14 @@
     <string name="refresh_data">Refresh data</string>
 
     <string name="delete">Delete</string>
+
+    <string name="about_developer">About developer</string>
+    <string name="about_city">About weather city</string>
+    <string name="about_temperature">About temperature</string>
+
+    <string name="action_settings">Settings</string>
+
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
 
 </resources>
